denest handle_dcc and make everything more readable

+++ kirc.h
@@ -58,7 +58,7 @@ static char *olog = NULL;       /* chat log path */
 static char *inic = NULL;       /* additional server command */
 static int cmds = 0;            /* indicates additional server commands */
 static char cbuf[CBUF_SIZ];     /* additional stdin server commands */
-static int ipv6 = 0;
+static short ipv6 = 0;
 
 #define htonll(x) ((1==htonl(1)) ? (x) : (((uint64_t)htonl((x) & 0xFFFFFFFFUL)) << 32) | htonl((uint32_t)((x) >> 32)))
 #define buffer_position_move(l, dest, src, size) memmove(l->buf + l->posb + dest, l->buf + l->posb + src, size)
